+++ src/components/Dropdown/components/menu/menu.jsx
@@ -4,6 +4,7 @@ import { components } from "react-select";
 import styles from "./menu.module.scss";
 
 const Menu = ({ children, Renderer, selectProps, dropdownMenuWrapperClassName, ...props }) => {
+  const rendererProps = { children, selectProps, ...props };
   const withFixedPosition =
     selectProps?.selectProps?.insideOverflowContainer || selectProps?.selectProps?.insideOverflowWithTransformContainer;
   return (
@@ -17,7 +18,7 @@ const Menu = ({ children, Renderer, selectProps, dropdownMenuWrapperClassName, .
         dropdownMenuWrapperClassName
       )}
     >
-      {Renderer && Renderer(props)}
+      {Renderer && Renderer(rendererProps)}
       {!Renderer && children}
     </components.Menu>
   );