+++ src/components/List/List.tsx
@@ -10,7 +10,14 @@ import { ListTitleProps } from "../ListTitle/ListTitle";
 import { ListWrapperComponentStringType, ListWrapperComponentType } from "./ListConstants";
 import { ComponentDefaultTestId, getTestId } from "../../tests/test-ids-utils";
 import { ListContext } from "./utils/ListContext";
-import { getListItemComponentType, getListItemIdByIndex, getListItemIndexById, useListId } from "./utils/ListUtils";
+import {
+  getListItemComponentType,
+  getListItemIdByIndex,
+  getListItemIndexById,
+  getNextListItemIndex,
+  getPrevListItemIndex,
+  useListId
+} from "./utils/ListUtils";
 import styles from "./List.module.scss";
 
 export interface ListProps extends VibeComponentProps {
@@ -75,9 +82,9 @@ const List: VibeComponent<ListProps> & {
         let overrideFocusIndex = undefined;
 
         if (isDownKey && focusIndex + 1 < childrenRefs.current.length) {
-          overrideFocusIndex = focusIndex + 1;
+          overrideFocusIndex = getNextListItemIndex(focusIndex, childrenRefs);
         } else if (isUpKey && focusIndex > 0) {
-          overrideFocusIndex = focusIndex - 1;
+          overrideFocusIndex = getPrevListItemIndex(focusIndex, childrenRefs);
         }
         if (overrideFocusIndex !== undefined) {
           updateFocusedItem(getListItemIdByIndex(childrenRefs, overrideFocusIndex));