+++ src/components/Menu/Menu/utils/utils.ts
@@ -4,3 +4,7 @@ export function isMenuChildSelectable(child: ReactElement): boolean {
   // @ts-ignore
   return !!child.type.isSelectable && !child.props.disabled;
 }
+
+export const generateMenuItemId = (menuId: string, child: ReactElement, index: number) => {
+  return (child.props as { id: string }).id || `${menuId}-item-${index}`;
+};