+++ src/components/index.js
@@ -31,6 +31,9 @@ export { default as Label } from "./Label/Label";
 export { default as Checkbox } from "./Checkbox/Checkbox";
 export { default as DialogContentContainer } from "./DialogContentContainer/DialogContentContainer";
 export { default as Dropdown } from "./Dropdown/Dropdown";
+export { default as DropdownMenu } from "./Dropdown/components/menu/menu";
+export { default as DropdownOption } from "./Dropdown/components/option/option";
+export { default as DropdownSingleValue } from "./Dropdown/components/singleValue/singleValue";
 export { default as Toast } from "./Toast/Toast";
 export { default as ToastButton } from "./Toast/ToastButton/ToastButton";
 export { default as ToastLink } from "./Toast/ToastLink/ToastLink";
@@ -87,7 +90,6 @@ export { default as ModalFooter } from "./Modal/ModalFooter/ModalFooter";
 export { default as ModalFooterButtons } from "./Modal/ModalFooter/ModalFooterButtons/ModalFooterButtons";
 export { default as Box } from "./Box/Box";
 export { default as Text } from "./Text/Text";
-export { default as Title } from "./Title/Title";
 export {
   useGridKeyboardNavigationContext,
   GridKeyboardNavigationContext