+++ webpack/published-js-components.js
@@ -2,6 +2,9 @@ const publishedJSComponents = {
   ResponsiveList: "components/ResponsiveList/ResponsiveList",
   MultiStepIndicator: "components/MultiStepIndicator/MultiStepIndicator",
   Dropdown: "components/Dropdown/Dropdown",
+  DropdownMenu: "components/Dropdown/components/menu/menu",
+  DropdownOption: "components/Dropdown/components/option/option",
+  DropdownSingleValue: "components/Dropdown/components/singleValue/singleValue",
   GridKeyboardNavigationContext: "components/GridKeyboardNavigationContext/GridKeyboardNavigationContext",
   useGridKeyboardNavigationContext: "components/GridKeyboardNavigationContext/GridKeyboardNavigationContext",
   allIcons: "components/Icon/Icons"